LayoutTests/platform/chromium-mac/plugins/window-open-expected.txt

CONSOLE MESSAGE: line 1: PLUGIN: WINDOW OPEN SUCCESS

This tests that window.open invoked by a plugin via NPN_Invoke without a javascript context succeeds.
